Fix security policy when changing security to "sensitive"
Previously we only applied security policy on new issues, that is
if you create an issue with security: none, then edit it to set
security:sensitive, it wouldn't apply a policy. This is now fixed.
Note: we still don't delete the policy when switching from
security:* to security:none, in that case users must remove the
custom security policy themself. There is simply no sane way to
implement the removal of policies. We have tried many strategies and
every implementation has dangerous edge cases.

Change-Id: Ia78e3ead4e85a84d77e390cf433be94a12e83f87